+++ R/utilities.R
@@ -42,10 +42,13 @@ na_if_empty <- function(item) {
 #'
 #' @examples
 #' \dontrun{
-#' idaifield_docs <- get_idaifield_docs(projectname = "rtest",
-#' connection = connect_idaifield(serverip = "127.0.0.1",
-#' user = "R",
-#' pwd = "password"))
+#' idaifield_docs <- get_idaifield_docs(
+#'   connection = connect_idaifield(
+#'     serverip = "localhost",
+#'     project = "rtest",
+#'     pwd = "password"
+#'   )
+#' )
 #'
 #' check_if_idaifield(idaifield_docs)
 #' }
@@ -213,34 +216,6 @@ name_docs_list <- function(docs) {
   return(docs)
 }
 
-
-#' TEMPORARY FUNCTION to warn about project(name)
-#'
-#' @param project connection$project
-#' @param fail should there be an error?
-#'
-#' @return TRUE/FALSE
-#'
-#' @keywords internal
-warn_for_project <- function(project = NULL, fail = FALSE) {
-  if (!is.null(project)) {
-    message <- paste("Please note: In the future, the project always",
-                     "has to be supplied to `connect_idaifield()`.",
-                     "Handing the project(name) as an argument to other",
-                     "functions will be deprecated in one of the next",
-                     "versions." )
-    warning(message)
-    if (fail) {
-      stop("Please supply a project to `connect_idaifield()`.")
-    } else {
-      return(TRUE)
-    }
-  }
-}
-
-
-
-
 #' Reduces a (relations) list to its first element and warns accordingly.
 #'
 #' @param resource The resource to be checked
@@ -352,4 +327,17 @@ find_named_list <- function(target_name, nested_list) {
   }
 
   return(NULL)  # Return NULL if no match found
+}
+
+#' Check if object is a "idf_connection_settings" and stop if not.
+#'
+#' @param x an object
+#'
+#' @return nothing
+#'
+#' @keywords internal
+stop_if_not_idf_connection_settings <- function(x) {
+  if (!inherits(x, "idf_connection_settings")) {
+    stop(paste0("'", deparse(substitute(x)), "' is not an 'idf_connection_settings'-object as returned by `connect_idaifield()`."))
+  }
 }