+++ tests/testthat/test-connect_idaifield.R
@@ -4,39 +4,16 @@ connection <- suppressMessages(connect_idaifield(serverip = "127.0.0.1",
                                                  ping = FALSE))
 ping <- suppressWarnings(idf_ping(connection))
 
-# Test that version is coerced to numeric
-test_that("version is coerced to numeric", {
-  connection <- suppressWarnings(connect_idaifield(version = "3",
-                                                   pwd = "hallo",
-                                                   project = "rtest",
-                                                   ping = FALSE))
-  expect_true(grepl("3001", connection$settings$base_url))
-  expect_is(connection$settings$base_url, "character")
-  expect_is(connection$status, "logical")
-  expect_is(connection$project, "character")
-})
-
-# Test that version less than 3 is set to 2
-test_that("version less than 3 is set to 2", {
-  connection <- suppressWarnings(connect_idaifield(version = 1,
-                                                   pwd = "hallo",
-                                                   project = "rtest",
-                                                   ping = FALSE))
-  expect_true(grepl("3000", connection$settings$base_url))
-})
-
 # Test that project argument is correctly assigned
 test_that("project argument is correctly assigned", {
-  connection <- suppressWarnings(connect_idaifield(version = 3,
-                                                   project = "rtest",
+  connection <- suppressWarnings(connect_idaifield(project = "rtest",
                                                    pwd = "hallo",
                                                    ping = FALSE))
   expect_equal(connection$project, "rtest")
 })
 
-test_that("warning is issued when no project is supplied", {
-  expect_warning(connect_idaifield(version = 3,
-                                   project = NULL,
+test_that("Fails when 'projet' is not set.", {
+  expect_error(connect_idaifield(project = NULL,
                                    pwd = "hallo",
                                    ping = FALSE),
                  "project")
@@ -44,12 +21,10 @@ test_that("warning is issued when no project is supplied", {
 
 # Test that ping argument is correctly assigned
 test_that("ping argument is correctly assigned", {
-  connection1 <- suppressWarnings(connect_idaifield(version = 3,
-                                                    pwd = "hallo",
+  connection1 <- suppressWarnings(connect_idaifield(pwd = "hallo",
                                                     project = "rtest",
                                                     ping = TRUE))
-  connection2 <- suppressWarnings(connect_idaifield(version = 3,
-                                                    pwd = "hallo",
+  connection2 <- suppressWarnings(connect_idaifield(pwd = "hallo",
                                                     project = "rtest",
                                                     ping = FALSE))
 
@@ -63,48 +38,11 @@ test_that("creates class 'idf_connection_settings'", {
   expect_true(inherits(connection, "idf_connection_settings"))
 })
 
-test_that("error for non-valid version number", {
-  expect_error(connect_idaifield(version = "test",
-                                 project = "rtest",
-                                 ping = ping),
-               "valid")
-})
-
-test_that("port 3001 for version 3", {
-  connection <- suppressMessages(connect_idaifield(version = 3,
-                                                   pwd = "hallo",
-                                                   project = "rtest",
-                                                   ping = ping))
-  expect_true(grepl("3001", connection$settings$base_url))
-})
-
-test_that("port 3001 for version 3, works with '3'", {
-  connection <- suppressMessages(connect_idaifield(version = "3",
-                                                   pwd = "hallo",
-                                                   project = "rtest",
-                                                   ping = ping))
-  expect_true(grepl("3001", connection$settings$base_url))
-})
-test_that("port 3001 for version 4", {
-  connection <- suppressMessages(connect_idaifield(version = 4,
-                                                   pwd = "hallo",
-                                                   project = "rtest",
-                                                   ping = ping))
-  expect_true(grepl("3001", connection$settings$base_url))
-})
-
-test_that("port 3000 for version 2", {
-  connection <- suppressMessages(connect_idaifield(version = "2",
-                                                   pwd = "hallo",
-                                                   project = "rtest",
-                                                   ping = FALSE))
-  expect_true(grepl("3000", connection$settings$base_url))
-})
-
 test_that("error, no valid ip", {
   expect_error(connect_idaifield(serverip = NULL, project = "rtest", ping = TRUE))
   expect_error(connect_idaifield(serverip = "klotz", project = "rtest", ping = TRUE))
   expect_error(connect_idaifield(serverip = 123, project = "rtest", ping = TRUE))
+  expect_error(connect_idaifield(serverip = "field.idai.world", project = "rtest", ping = TRUE))
 })
 
 test_that("auth object attached", {
@@ -129,3 +67,14 @@ test_that("error, no connection", {
 })
 
 
+test_that("adds correct limit to the query (needs updating when test data changes)", {
+  conn <- connect_idaifield(pwd = "hallo", project = "rtest", ping = TRUE)
+  query <- '{ "selector": { "resource.processor": ["Anna Allgemeinperson"] } }'
+  new_query <- add_limit_to_query(query, conn)
+  expect_true(grepl("75", new_query))
+})
+
+test_that("adds correct limit to the query (needs updating when test data changes)", {
+  query <- '{ "selector": { "resource.processor": ["Anna Allgemeinperson"] } }'
+  expect_error(add_limit_to_query(query, list(hello = "123")), "idf_connection_settings")
+})