+++ NEWS.md
@@ -1,3 +1,17 @@
+# idaifieldR 0.3.7 _2026-03-XX_
+* reworked `connect_idaifield()` (and affected): 
+  * 'project' parameter is now required. 
+  * 'version' and 'user' parameter removed. 
+  * 'localhost' new default for server.
+  * And the deprecated 'project' / 'projectname' parameter in many other functions is now finally removed.
+  * Substituted check for project parameter with structure check on connection parameter. If `connect_idaifield()` produced if, we just assume it's correct.
+* Improvement: `options(digits=20)` is reset to previous value after exiting `get_idaifield_docs()`. Sorry.
+* Bug: `add_limit_to_query()` now validates the updates query again. Unlike before.
+
+
+# idaifieldR 0.3.6 _2025-11-29_
+* Quick and dirty fix to get around the default limit of 25 for the `_find` enpoint of PouchDB since Field Desktop v3.6.0
+
 # idaifieldR 0.3.5 _2025-03-08_
 * Some minor fixes: 
   * checks if project exists in `simplify_idaifield()`